+++ CashReceipt.txt
@@ -4,8 +4,8 @@
               12. MILKYWAY Galaxy/ Earth
                 Tel: 123-456-70000
 
-Cashier: #1520                               DATE:2023-03-11
-                                               TIME:15:22:20
+Cashier: #1520                               DATE:2023-03-12
+                                               TIME:00:04:17
 DISCOUNT CARD: 1234
 ------------------------------------------------------------
  QTY                 DESCRIPTION          PRICE       TOTAL
