+++ src/main/java/com/clevertec/CheckRunner/model/DiscountCard.java
@@ -1,0 +1,19 @@
+package com.clevertec.CheckRunner.model;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "discount_card")
+public class DiscountCard {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    private Integer numberCard;
+    private Double discountPercent;
+}
