--- src/main/java/com/clevertec/CheckRunner/models/DiscountCard.java
@@ -1,19 +1,0 @@
-package com.clevertec.CheckRunner.models;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "discount_card")
-public class DiscountCard {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-    private Integer numberCard;
-    private Double discountPercent;
-}
