+++ src/main/java/com/clevertec/CheckRunner/repositorie/ProductRepositories.java
@@ -1,0 +1,9 @@
+package com.clevertec.CheckRunner.repositorie;
+
+import com.clevertec.CheckRunner.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface ProductRepositories extends JpaRepository<Product,Long> {
+
+}
