--- src/main/java/com/clevertec/CheckRunner/repositories/ProductRepositories.java
@@ -1,9 +1,0 @@
-package com.clevertec.CheckRunner.repositories;
-
-import com.clevertec.CheckRunner.models.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface ProductRepositories extends JpaRepository<Product,Long> {
-
-}
