+++ src/main/java/com/clevertec/CheckRunner/service/ProductService.java
@@ -1,0 +1,18 @@
+package com.clevertec.CheckRunner.service;
+
+import com.clevertec.CheckRunner.model.Product;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ProductService {
+    List<Product> findAllProducts();
+
+    Optional<Product> findById(long id);
+
+    boolean saveProduct(Product product);
+
+    boolean deleteProduct(Long id);
+
+    boolean updateProduct(Long id, Product product);
+}
