+++ src/main/java/com/clevertec/CheckRunner/util/mapper/BasketToString.java
@@ -1,0 +1,9 @@
+package com.clevertec.CheckRunner.util.mapper;
+
+import com.clevertec.CheckRunner.model.Basket;
+
+public interface BasketToString {
+
+    String printCheck(Basket basket);
+}
+
