+++ src/main/java/com/clevertec/CheckRunner/util/write/BasketWrite.java
@@ -1,0 +1,8 @@
+package com.clevertec.CheckRunner.util.write;
+
+import com.clevertec.CheckRunner.model.Basket;
+
+public interface BasketWrite {
+
+    void write(Basket basket);
+}
