--- src/main/java/com/clevertec/CheckRunner/utils/mapper/BasketToString.java
@@ -1,9 +1,0 @@
-package com.clevertec.CheckRunner.utils.mapper;
-
-import com.clevertec.CheckRunner.models.Basket;
-
-public interface BasketToString {
-
-    String printCheck(Basket basket);
-}
-
