--- src/main/java/com/clevertec/CheckRunner/utils/write/BasketWrite.java
@@ -1,8 +1,0 @@
-package com.clevertec.CheckRunner.utils.write;
-
-import com.clevertec.CheckRunner.models.Basket;
-
-public interface BasketWrite {
-
-    void write(Basket basket);
-}
