+++ src/test/java/com/clevertec/CheckRunner/service/impl/LRUCacheImplTest.java
@@ -1,0 +1,27 @@
+package com.clevertec.CheckRunner.service.impl;
+
+import org.junit.jupiter.api.Test;
+
+class LRUCacheImplTest {
+
+    @Test
+    void get() {
+
+    }
+
+    @Test
+    void put() {
+    }
+
+    @Test
+    void getSilent() {
+    }
+
+    @Test
+    void remove() {
+    }
+
+    @Test
+    void size() {
+    }
+}
