--- src/test/java/com/clevertec/CheckRunner/services/impl/LRUCacheImplTest.java
@@ -1,27 +1,0 @@
-package com.clevertec.CheckRunner.services.impl;
-
-import org.junit.jupiter.api.Test;
-
-class LRUCacheImplTest {
-
-    @Test
-    void get() {
-
-    }
-
-    @Test
-    void put() {
-    }
-
-    @Test
-    void getSilent() {
-    }
-
-    @Test
-    void remove() {
-    }
-
-    @Test
-    void size() {
-    }
-}
