+++ src/test/java/com/clevertec/CheckRunner/util/TestBuilder.java
@@ -1,0 +1,5 @@
+package com.clevertec.CheckRunner.util;
+
+public interface TestBuilder<T> {
+    T build();
+}
