--- src/test/java/com/clevertec/CheckRunner/utils/TestBuilder.java
@@ -1,5 +1,0 @@
-package com.clevertec.CheckRunner.utils;
-
-public interface TestBuilder<T> {
-    T build();
-}
